# src/main/java/com/example/bankcards/dto/response/JwtResponse.java
package com.example.bankcards.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class JwtResponse {
    private String token;
    private String username;
}