# src/main/java/com/example/bankcards/dto/response/UserResponse.java
package com.example.bankcards.dto.response;

import com.example.bankcards.entity.enums.Role;
import lombok.Builder;
import lombok.Data;

import java.time.LocalDateTime;

@Data
@Builder
public class UserResponse {
    private Long id;
    private String username;
    private String email;
    private Role role;
    private LocalDateTime createdAt;
}